requirements.txt
Flask==2.3.3
ultralytics==8.0.196
Pillow==10.0.0
gunicorn==20.1.0
opencv-python-headless==4.8.0.76